Rebase Client class constructor #108

+++ libs/Communicator/Client.cpp
@@ -12,7 +12,7 @@
 
 using namespace communicator_lib;
 
-Client::Client(std::string address, long port)
+Client::Client(std::string address, unsigned short port)
 {
     this->_address = address;
     this->_port = port;
